Notes
Here is the code I have been working on the past couple of days. I basically have the layout of the website set and the code organized. I currently have the CSS in the main file, but might been moved to it own later on. The carousal is working and so the the button the bottom to bring you back to the top of the page. The code is ready and label, so it will be easy to add in new features and information from the client.

Thanks Alexis